feat: Add AddTask component for creating new tasks
This commit adds a new component called AddTask, which allows users to create new tasks. The component includes form fields for entering the task name, description, time, and category. When the form is submitted, the task is added to the database using an HTTP POST request. This feature enhances the functionality of the application by providing a way for users to easily create and manage their tasks.